refactor: finance section
