+++ src/sections/finance/components/finance-withdraw.tsx
@@ -10,7 +10,7 @@ import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 
 // LOCAL IMPORTS
-import NeonPaper from '@src/sections/publication/NeonPaperContainer';
+import NeonPaper from '@src/sections/publication/components/neon-paper-container.tsx';
 import FinanceDialogsActions from '@src/sections/finance/components/finance-dialogs-actions';
 import TextMaxLine from '@src/components/text-max-line';
 import { formatBalanceNumber } from '@src/utils/format-number';
